+++ SOFTWARE/3D_Printer/Core/Src/Functions.c
@@ -1,11 +1,11 @@
 #include "Functions.h"
 
-void DelayMicrosecond(TIM_HandleTypeDef *timer, uint8_t time)
+void DelayMicrosecond(TIM_HandleTypeDef *timer, uint16_t time)
 {
     __HAL_TIM_SET_COUNTER(timer,0);
 	while (__HAL_TIM_GET_COUNTER(timer) < time);
 }
-void SelectDriver(uint8_t axis)
+void SelectDriver(uint16_t axis)
 {
     switch(axis)
     {
@@ -28,7 +28,7 @@ void SelectDriver(uint8_t axis)
     }
 }
 
-double GetTemperature(uint8_t adc, uint16_t value)
+double GetTemperature(uint16_t adc, uint16_t value)
 {
     double resistance, val;
     switch (adc)
